# app-backend/docs/swagger.yaml
basePath: /
definitions:
...
  users.BankWithdrawalRequest:
    properties:
      accountNumber:
        type: string
      amount:
        description: cNGN, which is Naira 1:1
        type: string
      bankCode:
        type: string
      bankName:
        type: string
      commit:
        type: integer
      messages:
        items:
          type: string
        type: array
      signatureRequired:
        type: integer
      transaction:
        type: string
      transactionId:
        type: string
      transactionSignature:
        type: string
      withdrawalId:
        type: string
    type: object
...
  users.StablerailProfile:
    properties:
      enabled:
        type: boolean
      minimumWithdrawal:
        type: string
      onboarded:
        type: boolean
      onboardingStatus:
        description: the last onboarding request's status, if any
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Core wallet API for the Trovo platform — user accounts, crypto assets,
    deposits/withdrawals, swaps.
  title: Trovo Wallet API
  version: "1.0"
paths:
...
  /v1/tokenization/payouts:
    get:
      description: 'The proceeds payouts (dividends / interest) to the user''s wallets,
        newest first: paid ones with their transaction, and scheduled ones once a
        payout''s holder schedule is locked. Optional tokenizedAssetId narrows it
        to one asset.'
      parameters:
      - description: Tokenized asset id
        in: query
        name: tokenizedAssetId
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      summary: GET /v1/tokenization/payouts
      tags:
      - users
...
  /v1/users/stablerail/profile:
    get:
      description: Whether bank deposits and withdrawals are available, whether the
        user's BVN is verified for them, and the smallest bank withdrawal.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/users.StablerailProfile'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      summary: GET /v1/users/stablerail/profile
      tags:
      - users
  /v1/users/stablerail/withdraw:
    post:
      consumes:
      - application/json
      description: 'Withdraws cNGN from the wallet to a Nigerian bank account. Like
        a payment it takes two calls: the first (amount, accountNumber, bankCode)
        returns the transaction to sign; the second (the same body plus transaction
        and transactionSignature, or commit for shared wallets) submits it. The cNGN
        goes to the user''s Stablerail wallet, which then pays the bank account; follow
        it with GET /v1/users/stablerail/withdrawals.'
      parameters:
      - description: Withdrawal
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/users.BankWithdrawalRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/users.BankWithdrawalRequest'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "503":
          description: Service Unavailable
          schema:
            additionalProperties: true
            type: object
      summary: POST /v1/users/stablerail/withdraw
      tags:
      - users
  /v1/users/stablerail/withdrawals:
    get:
      description: 'The user''s bank withdrawals, newest first, with their status:
        DEPOSITING, DEPOSITED, REQUESTING, DEPOSIT_FAILED, REQUEST_FAILED, then Stablerail''s
        (pending, processing, bank_verification, transfer_pending, transfer_confirmed,
        payout_pending, completed, failed, cancelled).'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      summary: GET /v1/users/stablerail/withdrawals
      tags:
      - users
...
securityDefinitions:
  BearerAuth:
    description: JWT bearer token used by the internal "trovo-manager" (tm-api admin)
      routes. Send as "Bearer <token>". See internal/middleware/authentication_middleware.go's
      JwtTokenAuthMiddleware.
    in: header
    name: Authorization
    type: apiKey
  ServiceLinkApiKey:
    description: Static per-partner API key used by white-label "service link" integrations
      (and by tm-api for a few admin/server-to-server calls). Issued when a service
      link is provisioned; see internal/components/servicelinks. See internal/middleware/api_key_middleware.go.
    in: header
    name: X-TW-SERVICE-LINK-API-KEY
    type: apiKey
  SignatureAuth:
    description: 'Request-signing scheme used by app-web and app-mobile for end-user
      requests. The caller sends four headers together: X-TW-PUBLIC-KEY (the wallet
      address making the call), X-TW-SIGNER (the address whose private key signed
      the request), X-TW-TIMESTAMP (a fresh timestamp), and X-TW-SIGNATURE (a signature
      over signer+timestamp, and for POST/PUT/PATCH over the request body/URI). There
      is no bearer token: every request is independently signed. See internal/middleware/authentication_middleware.go.'
    in: header
    name: X-TW-SIGNATURE
    type: apiKey
swagger: "2.0"